requirements.txt
streamlit
pandas
numpy
plotly
requests
yfinance
feedparser
joblib
python-dotenv
lxml
beautifulsoup4
